backend/app/requirements.txt
fastapi
uvicorn
pymongo
boto3
requests
python-dotenv
